Sets goban height and width to match svg.
Original script defaulted to always creating a 320 x 320 block diagram when rendered.
After this commit, we use the height and width attributes of the svg to set the correct
size on the rendered block diagram.  Because the goban has not rendered in Obsidian
yet, we update the sltxt2svg script to return the correct image sizes when it returns the xml,
and then we use those values.