feat: add the ability to use numbers >10 for move numbers and fixes offset bug for move #2 (#3)
Move Numbers > 10
We implement this by checking the row before parsing.
If the row appears to include move numbers, has spaces between
the intersections, and is not a metadata row for defining
lines and arrows, then instead of removing all spaces and splitting
by character, we can safely split the row by the space character.
This thereby allows multi-digit move numbers.

Offset Bug for Move #2
In order to fit a multi-digit numbers onto the stone, we need to
set the x-offset for larger numbers.  This inadvertently
fixed a bug whereby move number 2 always appeared offset to the left.